+++ service/resource/cloudformation/error.go
@@ -40,3 +40,24 @@ var wrongTypeError = microerror.New("wrong type")
 func IsWrongType(err error) bool {
 	return microerror.Cause(err) == wrongTypeError
 }
+
+var malformedAmazonAccountIDError = microerror.New("malformed amazon account ID")
+
+// IsMalformedAmazonAccountID asserts malformedAmazonAccountIDError.
+func IsMalformedAmazonAccountID(err error) bool {
+	return microerror.Cause(err) == malformedAmazonAccountIDError
+}
+
+var wrongAmazonAccountIDLengthError = microerror.New("wrong amazon account ID length")
+
+// IsWrongAmazonIDLength asserts wrongAmazonAccountIDLengthError.
+func IsWrongAmazonAccountIDLength(err error) bool {
+	return microerror.Cause(err) == wrongAmazonAccountIDLengthError
+}
+
+var emptyAmazonAccountIDError = microerror.New("empty amazon account ID")
+
+// IsEmptyAmazonAccountID asserts emptyAmazonAccountIDError.
+func IsEmptyAmazonAccountID(err error) bool {
+	return microerror.Cause(err) == emptyAmazonAccountIDError
+}
